refactor: Move TypeForm types to higher-level package.
