+++ src/main/java/org/opentripplanner/middleware/triptracker/TripSurveyUploadJob.java
@@ -9,8 +9,8 @@ import org.opentripplanner.middleware.connecteddataplatform.IntervalUploadFiles;
 import org.opentripplanner.middleware.connecteddataplatform.IntervalUploadJob;
 import org.opentripplanner.middleware.connecteddataplatform.ReportingInterval;
 import org.opentripplanner.middleware.models.TripSurveyUpload;
-import org.opentripplanner.middleware.models.typeform.Responses;
-import org.opentripplanner.middleware.models.typeform.Form;
+import org.opentripplanner.middleware.typeform.Responses;
+import org.opentripplanner.middleware.typeform.Form;
 import org.opentripplanner.middleware.persistence.Persistence;
 import org.opentripplanner.middleware.utils.DateTimeUtils;
 import org.opentripplanner.middleware.utils.FileUtils;
